Remove bukkit and spigot support.
Bukkit is in a weird state right now, no updates coming out. There
was some issues with the developers and mojang.

Spigot has forked off bukkit and seem to require manual building
right now, which would introduce a lot of new work to the cookbook
that I'm not really willing to do right now.

So I opted to remove support for both of them for now. If the
situation improves down the road we can always bring them back. I
think most of us working on the cookbook are only using vanilla
anyways.